requirements.txt
google-generativeai
requests
python-dotenv
flask
gunicorn
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Authlib
APScheduler
stripe
Flask-Limiter
cryptography
bleach
Flask-WTF
marshmallow==4.0.1
